Sexual 8bit music
Make sure javamod.jar is in the buildpath. I didn't make the music, I
mean, I can make 8bit music but I think it's too shite for the game. We
have to email the guy who made it if we want to use it so we'll do that
if you like it