Crosshair and more.
Added a basic crosshair to see where bullets will land, as well as some
tweaks to Npc spawning. Fixed bullets killing zombies a bit more, but
it's still not 100% Improved player death mechanism, and tried (and
failed) to add a death sound when player is killed by zombie - I shall
feex when I get a chance!